upstream lb: move all subset related code to extensions/
