+++ source/common/upstream/BUILD
@@ -258,26 +258,11 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "subset_lb_config_lib",
-    srcs = ["subset_lb_config.cc"],
-    hdrs = ["subset_lb_config.h"],
-    deps = [
-        "//envoy/upstream:load_balancer_interface",
-        "//envoy/upstream:upstream_interface",
-        "//source/common/config:utility_lib",
-        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/load_balancing_policies/common/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/load_balancing_policies/subset/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_library(
     name = "load_balancer_lib",
     srcs = ["load_balancer_impl.cc"],
     hdrs = ["load_balancer_impl.h"],
     deps = [
         ":scheduler_lib",
-        ":subset_lb_config_lib",
         "//envoy/common:random_generator_interface",
         "//envoy/runtime:runtime_interface",
         "//envoy/stats:stats_interface",
