+++ source/extensions/load_balancing_policies/subset/BUILD
@@ -14,6 +14,7 @@ envoy_cc_library(
     srcs = ["subset_lb.cc"],
     hdrs = ["subset_lb.h"],
     deps = [
+        ":subset_lb_config_lib",
         "//envoy/runtime:runtime_interface",
         "//envoy/upstream:load_balancer_interface",
         "//source/common/common:assert_lib",
@@ -21,11 +22,8 @@ envoy_cc_library(
         "//source/common/config:metadata_lib",
         "//source/common/protobuf",
         "//source/common/protobuf:utility_lib",
-        "//source/common/upstream:load_balancer_lib",
         "//source/common/upstream:upstream_lib",
         "//source/extensions/load_balancing_policies/common:factory_base",
-        "//source/extensions/load_balancing_policies/maglev:maglev_lb_lib",
-        "//source/extensions/load_balancing_policies/ring_hash:ring_hash_lb_lib",
         "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/load_balancing_policies/subset/v3:pkg_cc_proto",
@@ -45,5 +43,21 @@ envoy_cc_extension(
         "//source/common/common:minimal_logger_lib",
         "//source/common/upstream:load_balancer_factory_base_lib",
         "//source/common/upstream:load_balancer_lib",
+    ],
+)
+
+envoy_cc_extension(
+    name = "subset_lb_config_lib",
+    srcs = ["subset_lb_config.cc"],
+    hdrs = ["subset_lb_config.h"],
+    deps = [
+        "//envoy/upstream:load_balancer_interface",
+        "//envoy/upstream:upstream_interface",
+        "//source/common/config:utility_lib",
+        "//source/common/upstream:load_balancer_lib",
+        "//source/common/upstream:upstream_lib",
+        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/load_balancing_policies/common/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/load_balancing_policies/subset/v3:pkg_cc_proto",
     ],
 )
