+++ test/mocks/upstream/cluster_info.h
@@ -35,28 +35,6 @@ class FilterChainManager;
 
 namespace Upstream {
 
-class MockLoadBalancerSubsetInfo : public LoadBalancerSubsetInfo {
-public:
-  MockLoadBalancerSubsetInfo();
-  ~MockLoadBalancerSubsetInfo() override;
-
-  // Upstream::LoadBalancerSubsetInfo
-  MOCK_METHOD(bool, isEnabled, (), (const));
-  MOCK_METHOD(envoy::config::cluster::v3::Cluster::LbSubsetConfig::LbSubsetFallbackPolicy,
-              fallbackPolicy, (), (const));
-  MOCK_METHOD(envoy::config::cluster::v3::Cluster::LbSubsetConfig::LbSubsetMetadataFallbackPolicy,
-              metadataFallbackPolicy, (), (const));
-  MOCK_METHOD(const ProtobufWkt::Struct&, defaultSubset, (), (const));
-  MOCK_METHOD(const std::vector<SubsetSelectorPtr>&, subsetSelectors, (), (const));
-  MOCK_METHOD(bool, localityWeightAware, (), (const));
-  MOCK_METHOD(bool, scaleLocalityWeight, (), (const));
-  MOCK_METHOD(bool, panicModeAny, (), (const));
-  MOCK_METHOD(bool, listAsAny, (), (const));
-  MOCK_METHOD(bool, allowRedundantKeys, (), (const));
-
-  std::vector<SubsetSelectorPtr> subset_selectors_;
-};
-
 // While this mock class doesn't have any direct use in public Envoy tests, it's
 // useful for constructing tests of downstream private filters that use
 // ClusterTypedMetadata.
